Added boilerplate for missing components, moved components to components folder, and fixed export in State.js
