+++ tic-tac-toe/src/State.js
@@ -1,6 +1,4 @@
-const playing = Symbol('playing');
-const player1Won = Symbol('player1Won');
-const player2Won = Symbol('player2Won');
-const draw = Symbol('draw');
-
-export default {playing, player1Won, player2Won, draw};
+export const playing = Symbol('playing');
+export const player1Won = Symbol('player1Won');
+export const player2Won = Symbol('player2Won');
+export const draw = Symbol('draw');
