+++ tic-tac-toe/src/components/Footer.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import './../styles/Footer.css';
+
+function Footer(props) {
+
+}
+
+export default Footer;
