+++ tic-tac-toe/src/components/ResetButton.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import './../styles/ResetButton.css';
+
+function ResetButton(props) {
+
+}
+
+export default ResetButton;
