+++ tic-tac-toe/src/components/Tile.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import './styles/Tile.css';
+import './../styles/Tile.css';
 
 function Tile(props) {
   return (
